% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifnerror.R
\name{ifnerror}
\alias{ifnerror}
\title{If-else Wrapper for Errors}
\usage{
ifnerror(expr, noerror, error = "warn", silent = T)
}
\arguments{
\item{expr}{expression to test for error.}

\item{noerror}{expression to be called if no error.}

\item{error}{one of 'warn' or 'stop', passed to stopwarn.}

\item{silent}{logical, if TRUE (default) then error returned silently.}
}
\description{
Tests if an expression evaluates to an error, if so calls stopwarn
otherwise does whatever is given as argument.
}
\details{
Tests for an error or nerror (no error) from given expression. If an error is found
then either returns NULL with warning or breaks code. If nerror then returns the given expression.
}
\examples{
\dontrun{
ifnerror(1/2,print("Success"))
ifnerror(1/"a",print("Success"),"stop",silent = F)
}

}
\seealso{
\code{\link{stopwarn}}
}
